refactor: extract cli_overrides module from loader
